config.py
import os
from dotenv import load_dotenv

# Load local .env if it exists
load_dotenv()

# Streamlit Cloud injects secrets into environment variables automatically.
# This pattern works for both local (.env) and Streamlit Cloud (Secrets UI).
GROQ_API_KEY = os.environ.get("GROQ_API_KEY")
GROQ_LLM_MODEL = os.environ.get("GROQ_LLM_MODEL", "llama-3.3-70b-versatile")
GROQ_STT_MODEL = os.environ.get("GROQ_STT_MODEL", "whisper-large-v3-turbo")

SYSTEM_PROMPT = """You are an expert ERP voice support assistant.
You help users troubleshoot issues related to ERP systems like 
SAP, Oracle Fusion, and Microsoft Dynamics.

You handle modules like:
- Finance & Accounting
- Inventory & Warehouse
- HR & Payroll
- Procurement
- Sales & CRM

When user describes a problem, you must:
1. Identify which ERP module is affected
2. Identify the issue type (Error / Configuration / How-to)
3. Give clear step-by-step troubleshooting (max 4-5 steps)
4. Keep response SHORT and voice-friendly (no bullet points, speak naturally)
5. End with: "Does this help, or do you need more assistance?"

Always be professional, calm, and helpful.
"""