requirements.txt
streamlit==1.32.0
groq==0.5.0
gtts==2.5.1
soundfile==0.12.1
numpy==1.26.4
python-dotenv